# perf(es/minifier): Make name mangler parallel