--- a/crates/swc_ecma_minifier/src/pass/mangle_names/analyzer/mod.rs
+++ b/crates/swc_ecma_minifier/src/pass/mangle_names/analyzer/mod.rs
@@ -31,8 +31,13 @@ impl Analyzer {
             .map(|v| v.0)
             .collect::<FxHashSet<_>>();
         preserved_symbols.extend(unresolved.iter().cloned());
-        self.scope
-            .rename(&mut map, &Default::default(), preserved, &preserved_symbols);
+        self.scope.rename(
+            &mut map,
+            &Default::default(),
+            &Default::default(),
+            preserved,
+            &preserved_symbols,
+        );
 
         map
     }