--- a/crates/swc_ecma_minifier/src/pass/mangle_names/analyzer/mod.rs
+++ b/crates/swc_ecma_minifier/src/pass/mangle_names/analyzer/mod.rs
@@ -31,12 +31,16 @@ impl Analyzer {
             .map(|v| v.0)
             .collect::<FxHashSet<_>>();
         preserved_symbols.extend(unresolved.iter().cloned());
+
+        let cost = self.scope.rename_cost();
+
         self.scope.rename(
             &mut map,
             &Default::default(),
             &Default::default(),
             preserved,
             &preserved_symbols,
+            cost > 256,
         );
 
         map