--- a/crates/swc_ecma_minifier/src/pass/mangle_names/analyzer/scope.rs
+++ b/crates/swc_ecma_minifier/src/pass/mangle_names/analyzer/scope.rs
@@ -108,6 +108,12 @@ impl Scope {
             })
             .collect::<Vec<_>>();
 
+        if cfg!(debug_assertions) {
+            for (k, _) in iter.clone().into_iter().flatten() {
+                assert!(to.contains_key(&k), "Duplicate entry?");
+            }
+        }
+
         to.extend(iter.into_iter().flatten());
     }