--- a/crates/swc_ecma_minifier/src/pass/mangle_names/analyzer/scope.rs
+++ b/crates/swc_ecma_minifier/src/pass/mangle_names/analyzer/scope.rs
@@ -110,7 +110,7 @@ impl Scope {
 
         if cfg!(debug_assertions) {
             for (k, _) in iter.clone().into_iter().flatten() {
-                assert!(to.contains_key(&k), "Duplicate entry?");
+                assert!(!to.contains_key(&k), "Duplicate entry?");
             }
         }