--- a/crates/swc_ecma_minifier/src/pass/mangle_names/analyzer/scope.rs
+++ b/crates/swc_ecma_minifier/src/pass/mangle_names/analyzer/scope.rs
@@ -6,7 +6,7 @@ use rustc_hash::{FxHashMap, FxHashSet};
 use swc_atoms::{js_word, JsWord};
 use swc_common::{collections::AHashMap, util::take::Take};
 use swc_ecma_ast::Id;
-use tracing::debug;
+use tracing::{debug, error};
 
 use crate::util::base54;
 
@@ -109,8 +109,10 @@ impl Scope {
             .collect::<Vec<_>>();
 
         if cfg!(debug_assertions) {
-            for (k, _) in iter.clone().into_iter().flatten() {
-                assert!(!to.contains_key(&k), "Duplicate entry?");
+            for (k, v) in iter.clone().into_iter().flatten() {
+                if to.contains_key(&k) {
+                    error!("Duplicate entry?: {}{:?} => {}", k.0, k.1, v);
+                }
             }
         }